requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
httpx>=0.26.0
google-genai>=0.1.0
python-dotenv>=1.0.0